Separate app from server
